+++ lua/codecompanion/adapters/http/ollama/get_models.lua
@@ -43,9 +43,8 @@ local function fetch_async(adapter, opts)
   local url = adapter.env_replaced.url
   running = true
   _cached_models[url] = _cached_models[url] or {}
-  local headers = {
-    ["content-type"] = "application/json",
-  }
+
+  local headers = adapter_utils.set_env_vars(adapter, adapter.headers) or {}
 
   local auth_header = "Bearer "
   if adapter.env_replaced.authorization then
@@ -126,7 +125,7 @@ function M.choices(self, opts)
     log:error("Could not resolve Ollama adapter in the `choices` function")
     return {}
   end
-  adapter_utils.get_env_vars(adapter)
+  adapter_utils.get_env_vars(adapter, { timeout = config.adapters.opts.cmd_timeout })
   local url = adapter.env_replaced.url
   local is_uninitialised = _cached_models[url] == nil